requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0
sqlalchemy==2.0.23
pyodbc==5.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
smtplib-ssl==1.0.0
